+++ HISTORY.rst
@@ -5,6 +5,9 @@ Release History
 0.1.5 (Not yet released)
 ++++++++++++++++++
 - Fix the __all__ list so that you can `from pyinter import *`.
+- Change classes to new style classes (inheriting from object)
+- Added an `add()` function to the `IntervalSet` class. This will add (union) an Interval inplace. (similarly to `set().add()`)
+- The optional iterable of Interval objects passed when initialising an `IntervalSet` is now added Interval by Interval so that initialising with overlapping intervals works as expected.
 
 0.1.4 (2013-11-16)
 ++++++++++++++++++